Rename ert.shared.models to ert.run_models
