+++ src/ert/cli/model_factory.py
@@ -15,11 +15,13 @@ from ert.cli import (
     TEST_RUN_MODE,
 )
 from ert.parsing.config_errors import ConfigWarning
-from ert.shared.models.ensemble_experiment import EnsembleExperiment
-from ert.shared.models.ensemble_smoother import EnsembleSmoother
-from ert.shared.models.iterated_ensemble_smoother import IteratedEnsembleSmoother
-from ert.shared.models.multiple_data_assimilation import MultipleDataAssimilation
-from ert.shared.models.single_test_run import SingleTestRun
+from ert.run_models import (
+    EnsembleExperiment,
+    EnsembleSmoother,
+    IteratedEnsembleSmoother,
+    MultipleDataAssimilation,
+    SingleTestRun,
+)
 
 if TYPE_CHECKING:
     from ert.storage import StorageAccessor
