+++ src/ert/gui/simulation/ensemble_smoother_panel.py
@@ -12,8 +12,8 @@ from ert.gui.ertwidgets.models.ertmodel import get_runnable_realizations_mask
 from ert.gui.ertwidgets.models.targetcasemodel import TargetCaseModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert.libres_facade import LibresFacade
+from ert.run_models import EnsembleSmoother
 from ert.shared.ide.keywords.definitions import ProperNameArgument, RangeStringArgument
-from ert.shared.models import EnsembleSmoother
 
 from .simulation_config_panel import SimulationConfigPanel
