+++ src/ert/gui/simulation/multiple_data_assimilation_panel.py
@@ -16,12 +16,12 @@ from ert.gui.ertwidgets.models.targetcasemodel import TargetCaseModel
 from ert.gui.ertwidgets.models.valuemodel import ValueModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert.libres_facade import LibresFacade
+from ert.run_models import MultipleDataAssimilation
 from ert.shared.ide.keywords.definitions import (
     NumberListStringArgument,
     ProperNameFormatArgument,
     RangeStringArgument,
 )
-from ert.shared.models import MultipleDataAssimilation
 
 from .simulation_config_panel import SimulationConfigPanel
