+++ src/ert/gui/tools/load_results/load_results_panel.py
@@ -8,8 +8,8 @@ from ert.gui.ertwidgets.models.activerealizationsmodel import ActiveRealizations
 from ert.gui.ertwidgets.models.valuemodel import ValueModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert.libres_facade import LibresFacade
+from ert.run_models.base_run_model import _LogAggregration, captured_logs
 from ert.shared.ide.keywords.definitions import IntegerArgument, RangeStringArgument
-from ert.shared.models.base_run_model import _LogAggregration, captured_logs
 
 
 class LoadResultsPanel(QWidget):
