+++ src/ert/run_models/ensemble_experiment.py
@@ -10,9 +10,9 @@ import _ert_com_protocol
 from ert._c_wrappers.enkf import RunContext
 from ert._c_wrappers.enkf.enums import HookRuntime, RealizationStateEnum
 from ert.ensemble_evaluator import EvaluatorServerConfig
-from ert.shared.models import BaseRunModel
-from ert.shared.models.base_run_model import ErtRunError
 from ert.storage import EnsembleAccessor, StorageAccessor
+
+from .base_run_model import BaseRunModel, ErtRunError
 
 if TYPE_CHECKING:
     from ert._c_wrappers.enkf import EnKFMain, QueueConfig
