+++ src/ert/run_models/ensemble_smoother.py
@@ -8,8 +8,9 @@ from ert._c_wrappers.enkf import RunContext
 from ert._c_wrappers.enkf.enums import HookRuntime, RealizationStateEnum
 from ert.analysis import ErtAnalysisError
 from ert.ensemble_evaluator import EvaluatorServerConfig
-from ert.shared.models import BaseRunModel, ErtRunError
 from ert.storage import StorageAccessor
+
+from .base_run_model import BaseRunModel, ErtRunError
 
 if TYPE_CHECKING:
     from ert._c_wrappers.enkf import EnKFMain, QueueConfig
