+++ tests/unit_tests/cli/test_model_hook_order.py
@@ -2,7 +2,7 @@ from unittest.mock import ANY, MagicMock, call
 from uuid import UUID
 
 from ert._c_wrappers.enkf.enums import HookRuntime
-from ert.shared.models import (
+from ert.run_models import (
     EnsembleSmoother,
     IteratedEnsembleSmoother,
     MultipleDataAssimilation,
