+++ tests/unit_tests/ensemble_evaluator/test_evaluator_tracker.py
@@ -10,7 +10,7 @@ from ert.ensemble_evaluator import EvaluatorTracker, state
 from ert.ensemble_evaluator.config import EvaluatorServerConfig
 from ert.ensemble_evaluator.event import EndEvent, SnapshotUpdateEvent
 from ert.ensemble_evaluator.snapshot import PartialSnapshot, SnapshotBuilder, Step
-from ert.shared.models.base_run_model import BaseRunModel
+from ert.run_models import BaseRunModel
 
 
 def build_snapshot(real_list: Optional[List[str]] = None):
