+++ tests/unit_tests/shared/test_log_capture.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-from ert.shared.models.base_run_model import captured_logs
+from ert.run_models.base_run_model import captured_logs
 
 logger = logging.getLogger(__name__)
